move permission check inside the claim menu. potion permission close #49

+++ common/src/main/java/io/github/flemmli97/flan/gui/ClaimMenuScreenHandler.java
@@ -1,5 +1,7 @@
 package io.github.flemmli97.flan.gui;
 
+import io.github.flemmli97.flan.api.ClaimPermission;
+import io.github.flemmli97.flan.api.PermissionRegistry;
 import io.github.flemmli97.flan.claim.Claim;
 import io.github.flemmli97.flan.claim.ClaimStorage;
 import io.github.flemmli97.flan.claim.PermHelper;
@@ -23,7 +25,7 @@ public class ClaimMenuScreenHandler extends ServerOnlyScreenHandler {
     private final Claim claim;
 
     private ClaimMenuScreenHandler(int syncId, PlayerInventory playerInventory, Claim claim) {
-        super(syncId, playerInventory, 1);
+        super(syncId, playerInventory, 1, claim);
         this.claim = claim;
     }
 
@@ -44,6 +46,7 @@ public class ClaimMenuScreenHandler extends ServerOnlyScreenHandler {
 
     @Override
     protected void fillInventoryWith(PlayerEntity player, Inventory inv, Object... additionalData) {
+        Claim claim = (Claim) additionalData[0];
         for (int i = 0; i < 9; i++) {
             switch (i) {
                 case 0:
@@ -54,21 +57,29 @@ public class ClaimMenuScreenHandler extends ServerOnlyScreenHandler {
                 case 2:
                     ItemStack perm = new ItemStack(Items.BEACON);
                     perm.setCustomName(ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenMenuGlobal, Formatting.GOLD));
+                    if (player instanceof ServerPlayerEntity && !this.hasEditPerm(claim, (ServerPlayerEntity) player))
+                        ServerScreenHelper.addLore(perm, ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenNoPerm, Formatting.DARK_RED));
                     inv.setStack(i, perm);
                     break;
                 case 3:
                     ItemStack group = new ItemStack(Items.WRITABLE_BOOK);
                     group.setCustomName(ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenMenuGroup, Formatting.GOLD));
+                    if (player instanceof ServerPlayerEntity && !this.hasEditPerm(claim, (ServerPlayerEntity) player))
+                        ServerScreenHelper.addLore(group, ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenNoPerm, Formatting.DARK_RED));
                     inv.setStack(i, group);
                     break;
                 case 4:
                     ItemStack potions = new ItemStack(Items.POTION);
                     potions.setCustomName(ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenMenuPotion, Formatting.GOLD));
+                    if (player instanceof ServerPlayerEntity && !this.hasPerm(claim, (ServerPlayerEntity) player, PermissionRegistry.EDITPOTIONS))
+                        ServerScreenHelper.addLore(potions, ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenNoPerm, Formatting.DARK_RED));
                     inv.setStack(i, potions);
                     break;
                 case 8:
                     ItemStack delete = new ItemStack(Items.BARRIER);
                     delete.setCustomName(ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenMenuDelete, Formatting.RED));
+                    if (player instanceof ServerPlayerEntity && !this.hasPerm(claim, (ServerPlayerEntity) player, PermissionRegistry.EDITCLAIM))
+                        ServerScreenHelper.addLore(delete, ServerScreenHelper.coloredGuiText(ConfigHandler.lang.screenNoPerm, Formatting.DARK_RED));
                     inv.setStack(i, delete);
                     break;
                 default:
@@ -90,37 +101,60 @@ public class ClaimMenuScreenHandler extends ServerOnlyScreenHandler {
                 ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
                 break;
             case 2:
-                player.closeHandledScreen();
-                player.getServer().execute(() -> PermissionScreenHandler.openClaimMenu(player, this.claim, null));
-                ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                if (this.hasEditPerm(this.claim, player)) {
+                    player.closeHandledScreen();
+                    player.getServer().execute(() -> PermissionScreenHandler.openClaimMenu(player, this.claim, null));
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                } else
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
                 break;
             case 3:
-                player.closeHandledScreen();
-                player.getServer().execute(() -> GroupScreenHandler.openGroupMenu(player, this.claim));
-                ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                if (this.hasEditPerm(this.claim, player)) {
+                    player.closeHandledScreen();
+                    player.getServer().execute(() -> GroupScreenHandler.openGroupMenu(player, this.claim));
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                } else
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
                 break;
             case 4:
-                player.closeHandledScreen();
-                player.getServer().execute(() -> PotionEditScreenHandler.openPotionMenu(player, this.claim));
-                ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                if (this.hasPerm(this.claim, player, PermissionRegistry.EDITPOTIONS)) {
+                    player.closeHandledScreen();
+                    player.getServer().execute(() -> PotionEditScreenHandler.openPotionMenu(player, this.claim));
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.UI_BUTTON_CLICK, 1, 1f);
+                } else
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
                 break;
             case 8:
-                player.closeHandledScreen();
-                player.getServer().execute(() -> ConfirmScreenHandler.openConfirmScreen(player, (bool) -> {
-                    if (bool) {
-                        ClaimStorage storage = ClaimStorage.get(player.getServerWorld());
-                        storage.deleteClaim(this.claim, true, PlayerClaimData.get(player).getEditMode(), player.getServerWorld());
-                        player.closeHandledScreen();
-                        player.sendMessage(PermHelper.simpleColoredText(ConfigHandler.lang.deleteClaim, Formatting.RED), false);
-                        ServerScreenHelper.playSongToPlayer(player, SoundEvents.BLOCK_ANVIL_PLACE, 1, 1f);
-                    } else {
-                        player.closeHandledScreen();
-                        player.getServer().execute(() -> ClaimMenuScreenHandler.openClaimMenu(player, this.claim));
-                        ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
-                    }
-                }));
+                if (this.hasPerm(this.claim, player, PermissionRegistry.EDITCLAIM)) {
+                    player.closeHandledScreen();
+                    player.getServer().execute(() -> ConfirmScreenHandler.openConfirmScreen(player, (bool) -> {
+                        if (bool) {
+                            ClaimStorage storage = ClaimStorage.get(player.getServerWorld());
+                            storage.deleteClaim(this.claim, true, PlayerClaimData.get(player).getEditMode(), player.getServerWorld());
+                            player.closeHandledScreen();
+                            player.sendMessage(PermHelper.simpleColoredText(ConfigHandler.lang.deleteClaim, Formatting.RED), false);
+                            ServerScreenHelper.playSongToPlayer(player, SoundEvents.BLOCK_ANVIL_PLACE, 1, 1f);
+                        } else {
+                            player.closeHandledScreen();
+                            player.getServer().execute(() -> ClaimMenuScreenHandler.openClaimMenu(player, this.claim));
+                            ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
+                        }
+                    }));
+                } else
+                    ServerScreenHelper.playSongToPlayer(player, SoundEvents.ENTITY_VILLAGER_NO, 1, 1f);
                 break;
         }
         return true;
+    }
+
+    private boolean hasEditPerm(Claim claim, ServerPlayerEntity player) {
+        return ((claim.parentClaim() != null && claim.parentClaim().canInteract(player, PermissionRegistry.EDITPERMS, player.getBlockPos()))
+                || claim.canInteract(player, PermissionRegistry.EDITPERMS, player.getBlockPos()));
+    }
+
+    private boolean hasPerm(Claim claim, ServerPlayerEntity player, ClaimPermission perm) {
+        if (claim.parentClaim() != null)
+            return claim.parentClaim().canInteract(player, perm, player.getBlockPos());
+        return claim.canInteract(player, perm, player.getBlockPos());
     }
 }
